cast trimmed to int as gmdate requires an int not string

+++ php/class-media.php
@@ -2244,7 +2244,7 @@ class Media extends Settings_Component implements Setup {
 
 		foreach ( $logs as $signature => $log ) {
 			foreach ( $log as $time => $entry ) {
-				$readable_time                        = gmdate( get_option( 'date_format' ) . ' ' . get_option( 'time_format' ), trim( $time, '_' ) );
+				$readable_time                        = gmdate( get_option( 'date_format' ) . ' ' . get_option( 'time_format' ), (int) trim( $time, '_' ) );
 				$logs[ $signature ][ $readable_time ] = $entry;
 				unset( $logs[ $signature ][ $time ] );
 				if (
